requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiofiles==23.2.1
    # via aiogram
aiogram==3.12.0
    # via -r requirements.in
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via aiogram
aiosignal==1.3.1
    # via aiohttp
alembic==1.13.2
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
attrs==24.2.0
    # via aiohttp
cachetools==5.5.0
    # via -r requirements.in
certifi==2024.7.4
    # via aiogram
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
    # via sqlalchemy
idna==3.7
    # via yarl
magic-filter==1.0.12
    # via aiogram
mako==1.3.5
    # via alembic
markupsafe==2.1.5
    # via mako
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
psycopg==3.2.1
    # via -r requirements.in
psycopg-binary==3.2.1
    # via -r requirements.in
pydantic==2.8.2
    # via aiogram
pydantic-core==2.20.1
    # via pydantic
pyyaml==6.0.2
    # via -r requirements.in
sqlalchemy==2.0.32
    # via
    #   -r requirements.in
    #   alembic
typing-extensions==4.12.2
    # via
    #   aiogram
    #   alembic
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
yarl==1.9.4
    # via aiohttp